🔧 fix: relocate files & compress localStorage data
Close #31
